+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Add support for WoW 10.1.0
 * Add new icons from WoW 10.1.0 to icon picker
+* Add support for AddOn compartment
 
 ## [1.11.0] - 2023-03-22